ocr-api/requirements.txt
flask
pytesseract
opencv-python-headless
numpy
Pillow